chore: Extract subcase management to doctest/parts/subcase.h
